Refactor KNOWN_LANGUAGES into {DEV,PROD}_LANGUAGES
Remove KNOWN_LANGUAGES in favor of DEV_LANGUAGES and PROD_LANGUAGES
which are lazily used in LANGUAGES_URL_MAP and LANGUAGES depending on
the value of DEV.

DEV_LANGUAGES are autogenerated from the contents of the locale/
directory in the root of the project.

PROD_LANGUAGES should be maintained manually.

In addition, the commit removes the RTL_LANGUAGES setting which wasn't
used anywhere.  Django's default value for LANGUAGES_BIDI is sufficient
and used by django.utils.translation.get_language_bidi.