Do not convert bytestrings to unicode.
Back when using MySQL database backend, we had to set utf8_bin collation
on Entity and Translation tables to make use of case-sensitive
comparisons. That caused Django to return bytestrings for any character
fields instead of unicode strings. We had to manually convert
bytestrings to unicode. Now that we switched to PostgreSQL, this is no
longer necessary.

For more details, see:
https://docs.djangoproject.com/en/1.8/ref/databases/#collation-settings