Fix admin by moving monkeypatches to appconfig.
Because Django now calls admin.autodiscover() in the AppConfig for the
admin contrib app, we had an issue where autodiscover() was getting
called before we monkeypatched the Django admin site object with our
own version that plays nice with session-csrf. Monkeypatching after
that is called means the replacement AdminSite hasn't discovered any
of our apps.

This fix moves the monkeypatching to pontoon.base's AppConfig, which
gets run before the admin app's AppConfig and patches the AdminSite
before autodiscover() is called.