Also commit files that aren't in the repository yet
We set stats to 0 after repository is updated and than reset them during
extract. But we never reset them for files that use assymetric file
format and aren't in the locale repository yet. Since we also rely on
stats when picking files to dump and commit, these files never ended
up in the repository. This is a regression and should land asap.