Bug 1191718: Track Entity.has_changed status in the database.
Instead of querying for translations every time we need to see if an entity has
changed for a particular locale, we instead store whether the entity has changed
in the database whenever translations are saved. This significantly speeds up
sync_projects by removing a ton of unnecessary database queries.

Since this replaces what Project.last_synced was used for, we remove that field.