Bug 1209159: Use AJAX to fetch main menu data
This changes the way we load project, locale, resource and stats data
for the main menu. Instead of loading data as part of the main server
request, we make separate AJAX requests for each project. Except for
the current project, which is still loaded as part of the main request
to make project resources/subpages/search immediately available.

PROS
- Decreases homepage main request (/) time by > 80%
- Decreases homepage total request (all) time by > 50%
- Decreases homepage fully loaded time (incl. new AJAX calls) time by > 20%
- Decreases homepage HTML size by > 80%

CONS
- Main menu is fully ready a few seconds after page load
- Possible: memory consumption