Fix bug 1217753: broken entity list
It was caused by:
https://github.com/mozilla/pontoon/commit/197bc8c70a810d9391bc9aaec158170bc9ca433c

Which implemented searching by string key by storing key in the data-key
attribute of each entity in the list. Since some keys also contain
strings and strings can contain HTML markup, we broke HTML of the
entity list.

This fix only puts keys that are not equal to source strings into the
attribute and also escapes keys.