Always close menu on click outside the menu
We close menus on clicks outside. In theory. But in practice, we use
stopPropagation in many event handlers, preventing the event from
bubbling to the top of the DOM, where we catch clicks outside.

Yes, stopping event propagation can be dangerous:
https://css-tricks.com/dangers-stopping-event-propagation/

This patch removes all event.stopPropagation() calls, which prevented
menus from closing. In most cases simply removing the call was enough.
I only had to slightly refactor event handlers for helpers.

In addtion to that, I also replaced " with ' in modified event handler
selectors.