Bug 1214411: Skip repo sync if locales have nothing to sync
I noticed the following sync messages in the logs:
"Synced translations for project firefox-aurora in locales ."

It occurs if repo changes, but none of the locale files is affected,
e.g. if a readme file changes in the repo.

As soon as we detect there aren't any changes in translation files,
we skip the repo sync.