Fix bug 1183999: Prioritize suggestions from other locales
This feature makes it possible for users to define a list of preferred
locales, which are displayed first (on top) in the locales tab in the
translate view.

A new settings page has been designed, accessible from the translate
view and the user profile page.

A quick settings menu has also been added to the translation view, near
the translation textarea. It includes quality checks and make suggestions
setting, along with the link to the general settings page.

Preferred and remaining locale counts are displated separately in the
locales tab. And a visual separator is displayed to distinct between
both suggestion groups in the locales tab.

A multiple locale selector widget has been added, which is used by the
settings page and the admin interface. Some transifex references have
also been removed.