Fix bug 1225305: Filter by translation time
Adds a new filter that allows for searching translations, submitted
within a time interval. Usese jQuery UI Date & Time Picker for basic
time range validation and a calendar-like date picker.

A few time range shortcuts are available for quick input. A local time
zone is used on the client side, which is then converted to UTC when
generating a URL parameters to be used on server, allowing URLs to be
shared and always return the same value.

Due to recent additions the filter menu has been slightly redesigned:
* It has become a modal window with fixed width and simplified selector,
  so no more funky background and title replacing the search field.
* Authors are listed at the bottom, so that the rest of the filters are
  always located in the same position. The menu can be scrolled if needed.
* Counts have been added to Translation Status filters.