Add translation count chart as a time range selector (#439)
We aggregate translation submission by minute every time a collection
of entities is requested and plot it on a chart which acts as a time
range selector. That way users don't need to shoot in the dark like they
would oftenvdo if date & time picker was only available in the form of a
calendar.

Also:
- On backend, increase time interval by a minute due to truncation.
- Sort authors by translation count and display count in the filter menu.