Fix adding new locales if source repository has changes
If repository containing source strings changed without changes to
source strings and a new locale was added to Pontoon, the following
sync used to obsolete all entities.

It happened because VCS was correctly detected as not having any entities,
but in DB all existing entities were detected and got obsolete in
update_entities().

We now correctly detect DB entities as not having any entities, too.