Fix bug 1305233: Properly match repository changes against locales (#541)
The old code checked if changed paths in repository start with paths to
locale directories without the trailing slash. E.g. if the changed path
was `kab/mozorg/internet-health/privacy-security.lang`, it would match
it with the Georgian locale directory `ka` and break.

The new code adds trailing slashes to locale directory paths, so it
checks if the path starts with `ka/` (no match) and then continues all
the way to `kab/` (match).

The bug affected locales with codes that could be partially overlapped
by other locale codes.