Bug 1344658: optimize authors and time range
Instead of updating a list of authors and the time range in the filter
menu on any event that can change them, we do it when the filter menu
is opened.

This is mostly noticeable when saving and deleting translations. Together
with 242de20c42663d8a0476cb1081deaae4df3b4312, those two tasks should be
at least 80% faster.

It also speeds up loading of the translate page and entities.