Fix bug 1377830: Fix selecting all entities
Previously, when selecting all entities for mass actions, we loaded PKs
of all entities that match the selected criteria for the string list.
We have to do this, because we use paging and not all entites are loaded
in the string list.

The problem with this approach is that some strings could have changed
state since they were loaded in the string list. For example, if only
missing strings were filtered, and we translated them, Select all would
not return such strings.

Now we concatenate those strings returned from the server with the
selected strings on the client side.