Bug 1382538: Store FTL translations without comments
That's because we need to serialize translations consistently on the
client when submitting translations and on the server during sync.

We could also change the code on the client and store translation
comments there, but that turns out to be harder to implement. Besides,
we don't really use translation comments.