Fix bug 1384846: Fix diff view for FTL strings
This worked for all formats except for FTL, because getSimplePreview()
returns the fallback value for them, which is passed as the second
argument.

For FTL strings we need to generate their presentation using their
source, which is passed as the first argument. Since the same argument
was passed for the base string and the translate string, the diff
was always empty.