Bug 1381852: Unapprove fuzzy translations on import (#679)
If an existing translation is made fuzzy in the VCS, we mark it as
fuzzy in the DB on next sync, but don't unapprove it. That is because
two separate instances of the same translation are passed to
bulk_update - the first one sets the fuzzy flag and the second one
unsets the approved flag. bulk_update only executes the first instance,
so now we properly set all flags on it and don't pass the second one.

Note: this only applies to translations that are made fuzzy in the
locale file. Most of the time a new source strings is created when
existing translation are made fuzzy on msgmerge. This case already
behaves properly.