Fix bug 1395490: fix KeyError on sync
When detecting changed resources, only take into account changes made
before the sync job started. Otherwise we might end up trying to sync
locales we didn't pull files for, which will lead to KeyError in
VCSResource constructor's for locale in locales loop. This is a
regression from bug 1383252.

That same error is also thrown when a new locale is enabled for a
project during sync - after the "pull phase". We fix it by checking
that all locales identified as needing sync have been pulled. This is
an old bug we never fixed.

Note: None of these two bugs is causing destructive changes: they break
the project sync which then completes on the next run. However, if new
source strings are added and the error occures, we hit bug 1383271.