Bug 1397725: Prefetch notifications project data
The average call to the /notifications page triggers 178 base_project
select DB queries, which makes the page slow. With this fix, the number
goes down to 2 and the page is much faster.

We use the same trick on the notifications menu. The impact is less
obvious, because we usually only load the latest 7 notifications.