Bug 1357361: Speed up contributors list (#706)
Fix bug 1357361: Speed up contributors list (#706)

Several speedups are included:

1. Do not recalculate the user role

2. Drop ability to exclude contributors

This option used to be valuable in the early days of Pontoon, when I
was the top contributor on Pontoon. I found it embarrassing to see my
name on top of the leaderboard.

3. Include unchanged translations in counts

Unchanged translations used to be excluded from Contributors dashboards
and Contributor profile stats and timeline. There's no good reason for
that. Unchanged translation is still a valid translation.

4. Faster user role detection

We create Translator Groups for each Locale and each ProjectLocale. They
are used to assign Translator permissions to users for either the entire
Locale or for specific projects within the locale.

When collecting user roles for the Contributors dashboards, we now only
check Locale Translator Groups, because:

* it speeds up user role detection from around 2.0 to around 0.1 seconds,
* it yields the same role for 96/100 Top Contributors,
* user role is not the critical information on Contributors dashboards,
* and the code is much more elegant now.

5. Faster user stats SQL query

This simplifies the SQL query to retrieve user stats and also makes it
almost 80% faster. Instead of annotating user objects with translation
counts using DB joins and SQL directly, we now only retrieve translation
counts per author and translation status from the DB. Then we use python
to calculate counts for each status and each user and assign those values
to user objects manually.

Additional benefit is that more specific filters (by time, project,
locale and localization) generate joins that are one level shallower
and as such, faster.