Fix bug 1399262: Reset screenshot on string navigation (#715)
If a string contains a link to a JPG or PNG file in its comment, we
render the linked image as a thumbnail in the source panel. If the
subsequently openned string does not have a comment, we keep the
screenshot visible instead of removing it. This patch fixes that.