Fix bug 1399166: Use X-channel repo as source (#714)
Update the script to create per-project en-US only repositories from
the cross-channel repository. It is already limited to localizable
files only, and it also includes files from the comm-* repositories.

We now also clean up target directory on a failed pull, so that it's
empty for a clone. We also prune all subdirectories in target
repository in case they get removed from source.