Fix bug 1229891 Add logging of changes in permissions of users.
Every change in a contributor's permissions should will be stored in PermissionChangeset model.
Unfortunately, request.user isn't available for The Django signals and because of that the logic is implemented in views (which at first may look less intuitive).
Currently, there are 2 views that are able to modify a contributor's permissions - the first one is only accessible for admin/superusers and the second is a locale's "Permissions" tab.