Fix bug 1407623 - New source input for short-term projects (#745)
* Fixes bug 1407623 - Add direct input for projects with no repository.
This is the first step of allowing a new source of strings for projects. It permits to create projects without any repositories, and to enter string in a textarea. This commit also enables string management, to edit source strings and comments.

* Re-styling for coherence with general code.

* Added a button to download all translations as CSV.

* Enable strings management only for projects with no repository.

* Enable adding new strings. Better UI and styles.

* Fix small issues.

* Add a data_source field to projects, choices being repository or database.
This allows to explicitely mark projects for which we want to manage strings in Pontoon as "database". It makes the logic of showing the manage strings page clearer.

This commit also adds a textarea on the project creation page, allowing users to send the first batch of strings directly during the creation, instead of having to go to the manage strings page afterwards. The links move around a bit as well.

* Better link names.

* Hide repositories when data source is database.

* Add option to remove a string. Rename path to "database" from "all". Fix conflict on migration.

* Do not run sync on database sourced projects.

* Restore section padding

* Move textarea label to top for consistency. Use semantic HTML. Remove en-US note (true for all projects)

* Data Source and Strings sections in a single line

* Use example in Placeholder to avoid duplicating the label

* Separate page action (New string) from navigation buttons

* I doubt this action is needed on this page and I dont like duplicating functionality. Feel free to revert if you disagree.

* Adding action to the Action button

* White space is nice

* use plural for new strings placeholder.