Fix bug 1382348: Mark placeables in FTL strings
Only references to other Messages and External Arguments are detected
as placeables, since FTL gives us this information and we're confident
when to mark them.

The Translate Toolkit placeables logic we use for other formats is not
usable here for various reasons:
* it's written in python and we build the UI in JS
* it yields more false positives since FTL messages contain more makup

It also has other drawbacks, so we should look into replacing it. See
bugs 1251186, 1368602.