Fix bug 1335331 - Enable users to change their email address (#793)
* Bug 1335331: added ability to change email address

* Bug 1335331: add username in read-only mode

* Bug 1335331: remove placeholders

* Bug 1335331: add email validation

* Bug 1335331: adjust contributor profile tests

* Bug 1335331: make profile-form submit more generic

* Bug 1335331: add logout after email change

* Bug 1335331: change email in tests

* Bug 1335331: add email warning message to css

* Bug 1335331: adjust contributor profile test email

* Bug 1335331: shorten the line to pass PEP8

* Bug 1335331: make code style consistent

* Bug 1335331: correct level of indentation

* Fix errors on settings page.
This commit removes the fancy javascript that allowed to save user profile without reloading the page. It was a big source of errors and complexity. The settings page is now much simpler, the custom home page being the only exception. This fixes the CSRF error that was happening as well when saving data by clicking the Save button.

* Show or hide warning using JS.
It was bad that the warning for a changed email would only appear on focus on the input. First it made it difficult to click the save button, as the layout was changed and thus the button moved. Second, it would show up even if the email was not changed.
This commit also moves CSS to the more specific settings.css file, and addresses a few review comments.

* Make message a help text instead, and always show it.

* Styling changes.