Bug 1431436: [FTL] Support select expressions in read-only display (#814)
Improve read-only presentation of FTL Messages to support select
expressions, including `PLATFORM()`. Select expressions can be either
in a value or in attributes. They can be surrounded by other elements
of any type.

FTL Message that contains any combination of elements of type
`TextElement`, `ExternalArgument`, `MessageReference` and
`SelectExpression` in their value or attribute or both are now
supportedcin read-only display.

There are two read-only displays the we use in the UI and have been
improved as part of this patch:

1. Simplified string preview:
* Now uses default variants of select expressions instead of first.
* Backend and frontend code are now unified.

2. Original string presentation.