Fix bug 1431436: [FTL] Support select expressions in editor (#819)
Improve FTL editor to support presenting and saving select expressions,
including `PLATFORM()`. Select expressions can be either in a value or
in attributes. They can be surrounded by other elements of any type.

FTL Messages that contain any combination of elements of type
`TextElement`, `ExternalArgument`, `MessageReference` and
`SelectExpression` in their value or attributes or both are now supported
in the editor.

Other improvements:
* Allow turning simple strings into complex string via source editor.
* Detect strings with parse errors as not supported in rich FTL editor
  and prevent switching to FTL editor until they are fixed.