Replace a list of all resources with an empty list
When entities for All Resources were requested, we sent a list of all
resource paths as the request parameter and queried entities in the DB.

If strings from All projects were displayed in the translate view, that
could lead to an error: "The number of GET/POST parameters exceeded
settings.DATA_UPLOAD_MAX_NUMBER_FIELDS".

The list of all resource paths also needed to be retrieved from the
server first, which wasn't very fast, especially when retrieving paths
for all locale projects.

From now on we use an empty list as a way of defining All Resources.