Fix bug 1219908: Search word by word (#888)
Change default search behaviour to search for each word separately and
show results that contain all of them. That's closer to the default
behaviour of search engines.

    Example: beta firefox

    Finds strings that contain "Firefox Beta".

To search for exact matches of multiple words, you can wrap them in
double quotes. If you wrap the entire string into double quotes, the
search behaviour is the same as in the past.

    Example: Open "New Window"

    Finds string "Open in a New Window".

If you want to search for strings that contain double quotes, you can
escape them with \.

    Example: <a href="%(url)s">

    Searches for `<a`, `href=`, `%(url)s` and `>`.

    Example: <a href=\"%(url)s\">

    Searches for `<a` and `href="%(url)s">`.