Fix bug 1453139: Prevent DoesNotExist for non-managers
The error was caused when User.can_translate() was called when checking
if Make Suggestions toggle should be available, and the Project with
slug "all-project" was passed as a parameter.

We now check for the visibility of the toggle for each string separately
and not on translate page load. Which is the right way of doing it,
because a user could have Translator permission for some strings
within the locale and not for some others.