Bug 1455246 - Improve pytest environment.
This commit moves the pytest configuration into a more discoverable file (pytest.ini), and changes it to run tests in a more standard manner. Running pytest now runs all django tests in the pontoon folder (and not tests in /tests) and reuses the database between runs. It also enables passing files and folders to the pytest command.