Bug 1458347: Move Fluent checks to backend (#943)
Remove frontend Fluent checks that are already included in
compare-locales and move the rest to pontoon.checks.libraries.pontoon.

Also, changed the logic for switching between the rich and source
editor (works both ways now). That's needed, because on editor switch
we also run checks (to avoid losing content if it's broken). A message
is displayed when switching to rich editor not possible, because the
string is only supported in source editor.

Other enhancements:
* Add border on top of warning overlay (to distinc from fluent editor)
* Catch Esc event if translation textarea not focused
* No period at the end of failed check messages

Code changes:
* Created a perform_checks() AJAX view
* Factored out renderFailedChecks() function
* Unified compare-locales and translate toolkit library interfaces to
  always take locale.code as argument (instead of locale or
  locale.code).
* Special logic for detecting unsaved changes no longer needed