Fix bug 1450505: [FTL] Re-enable Tab to copy from helpers
In complex FTL strings, multiple fields are present, and we use Tab to
navigate between them, as this is the browser default shortcut. We also
use Tab to navigate between helper suggestions (history, machinery,
locales), as many other CAT tools do.

This patch re-enables Tab for navigating helper suggestions for complex
FTL strings when the focus is in the last field. After the suggestion
is copied, we keep the focus in the last field to make moving to next
suggestion with subsequent Tabs easier.