Fix bug 1456846 - Fix docker dev environment. (#942)
* Fix bug 1456846 - Fix docker dev environment.

This commit ensures that:
 - the assets/ folder is always created
 - dependencies (Python and node) do not pollute the user's repository. They are now installed in a different folder inside the container
 - webpack takes resources from the right place, and puts bundles in the right folder

This looks like a small patch but it was actually fairly complex. The reason is that our Dockerfile is made for a production deployment, meaning that it builds all the resources it needs to run in place. That's fine. However, when we run locally for development, using docker-compose, we override the /app folder with the user's local folder. That means all resources built during `make build` are lost.
This patch puts those resources in a different folder, and then links to that folder from the user's repository. This allows to keep a clean production build process, while having a working dev environment, and a clean local folder.