Fix bug 1377969: Introduce unreviewed suggestions (#926)
This commit removes the Suggested translation status and introduces
Unreviewed suggestions. The benefit is that Unreviewed suggestions also
include suggestions submitted to Translated and Fuzzy strings, which
were previously inaccessible from dashboards and thus easy to miss.

Translations that were previously in a Suggested state are now treated
as Missing, which is what they are in the repository (and product), too.

Unreviewed suggestions are displayed on all dashboards and also
available as a filter. Since Unreviewed suggestions cannot be painted
on progress charts, we're adding a new sortable Unreviewed suggestions
column to dashboards, represented by a lightbulb, painted blue if
unreviewed suggestions are present.

We're also introducing a new calculate_stats management command, which
re-calculates statistics for all translated resources and corresponding
objects.

As part of this commit, we made a few other changes:
* Renamed translated_strings to unreviewed_strings.
* Exposed unreviewed_strings in the API.
* Upgraded to Font-Awesome 5.