Fix bug 1470633: Add locales to the bottom of the list (#981)
In multiple locale selector, we used to prepend locales to the list
when they were selected. Now we append them.

Affected pages are Settings, Project Admin and Project Notificaions.

The benefit is especially useful on the Setting page, where preferred
locales are selected for which the order is important. The one selected
first will now also be the one on top of the list in the Locales tab in
translate view.