Fix bug 1464802: Show priority + deadline in localization dashboard (#1004)
The most critical piece of information coming from the tags feature
is priority. It's only available in the tags tab, which as used very
rarely.

We're fixing this by exposing file priority in the Resources tab of the
Localization dashboard. The column only appears if at least one file is
assigned to a tag (which sets the priority).

Since it's touching the same code and serving a similar purpose, we're
also adding a deadline column to the localization dashboard. And,
factoring out the deadline and priority field code, so that it can be
reused easily.