Fix bug 1478464: Read-only locale admin (#1035)
This patch adds the ability to manage read-only locales in Pontoon
Admin. Since the locale selector is now wider, the entire Project Admin
now also uses a full-width layout2.

Parts of the code have been refactored to use as much template, CSS and
JS code as possible. The new locale selector widget shares many
similarities with the locale selector from the Settings and the
Notifications sending page, and the contributor selector from the
Permissions page.