Fix bug 1482154 - Add a demo project
This project will be used as a dynamic tutorial for new users. It's available
for all languages, and is the one project where we show the guided tour. It is
stored in our database and does not sync with any repositories.

Updates to the content of that project should be handled with new migrations,
potentially reusing code from the initial migration.

We intentionally avoided using existing views (e.g. for managing strings for
DB-based projects) to prevent breakage in the future.