Fix bug 1482154: Add a tutorial project (#1053)
The tutorial project will be used as a dynamic tutorial for new users.
It's available for all languages, and is the project we show the guided
tour on. It's stored in the DB, does not sync with any repository and
is a system project - hidden from dashboards.

Updates to the content of the project should be handled with new
migrations, potentially reusing code from the initial migration. We 
intentionally avoided using existing views (e.g. for managing strings
for DB-based projects) to prevent breakage in the future.

Tutorial and guided tour have been moved to a standalone pontoon.tour
Django app. We also used this opportunity for updating copy and look
& feel of the tour (including speeding up animations).