Fix bug 1378471: New Pontoon homepage (#1002)
The new Pontoon homepage has been designed and developed by @shkamaru
during Google Summer of Code 2018. It presents Pontoon as the place to
localize Mozilla projects, instead of an in-context localization tool
as it was the case in the past.

It explains the whys and hows of localization at Mozilla, introduces a
clear call to action to start contributing to Mozilla localization and
moves in-context localization demo to a separate page. It also links to
a guide tour and makes Sign in link more visible by moving it from the
profile menu to header.

It resides in a standalone Django app, pontoon.homepage.