Fix bug 1481766 - Show entity metadata. (#1051)
* Fix bug 1481766 - Show entity metadata.

This adds a new component to show the metadata of a selected entity. It also refactors the component and the Editor component into a new, high-level (redux-connected) component. This makes the logic a bit easier to test, reduces the number of redux-connected components, and makes more sense semantically.

Features of this patch:
- show entity original string
- show comment if any
- show sources (array or object)
- show context if any
- turn URLs into links
- show screenshots miniatures and a lightbox when clicked
- remove MAX_LENGTH from comments