Fix bug 1481175: Store active translation in the DB (#1072)
This patch introduces a Translations.active field. It is Boolean and
has a constraint to be True for a maximum of one unique combination of
(locale, entity, plural_form). Active translations are displayed in the
string list and as the first entry in the History tab.

The idea is to store active translations in the DB, instead of
calculating them on the fly, which is expensive. Hence, we're adding a
data migration to set Translation.active to True where needed for
existing translations. We also make sure that the active field is
updated properly when translations are changed. And finally, we're
making use of the active field instead of calculating it in the code.

For practical reasons, we're also fixing two other bugs:
* Bug 1477581: Unfuzzy on mass reject (and approve). These changes
  prevent translations from being in double state: fuzzy and approved
  or rejected.
* Bug 1397138: Adding suggestions must not unfuzzy strings.

The patch also removes the Translation.entity_document field and
related trigger and function, which were introduced earlier and aren't
used anymore.

It also makes code for unapproving, unfuzzying and rejecting
translations more consistent and maintainable.