Fix bug 1486516 - Continuous loading of entities.  (#1073)
Fix bug 1486516 - Continuous loading of entities.

This makes sure that the screen always takes 100% of the screen, and that scrolling is handled in panels, notably in the EntitiesList component. It also adds an infinite scroller so that next entities are loaded when a user goes near the bottom of the list (by default, within 250px from the bottom).