Use errors and warnings statuses only for strings in VCS
In practice that means approved and fuzzy strings only, instead of all
active strings (which also include unreviewed suggestions).

That's because only strings in VCS require special attention, i.e.
presentation on dashboards and the ability to be filtered, because they
could be breaking the product.

Errors and warnings in unreviewed suggestions are not only displayed
on dashboards and filterable (as unreviewed suggestions), but will also
show up if somebody tried to approve them (as failed quality checks).

That change also means error/warning numbers in Pontoon dashboards
should match those numbers on Elmo and web dashboards.

And yes, it also means that it's nice that we implemented Translation.active,
but we don't need it for checks.