Fix bug 1491027 - Filter on project when searching for tags. (#1087)
The lack of filtering on projects means that before this patch, finding the tag that was being concerned only took the slug into consideration. Thus, if two tags, under different projects, used the same slug, one of them would show up for all projects, leading to an incorrect listing of resources, and impossibility to manage that tag.