Fix bug 1486517 - Add support for plurals to Translate.Next. (#1096)
* Fix bug 1486517 - Add support for plurals to Translate.Next.

For entities that have a plural form, the History tab shows only translations for the selected plural form. The Editor allows a user to change the plural form, which updates the History and the content of the Editor. Plural examples are loaded from the backend, through the graphql API (yay! ).