Fix bug 1501168: Upload FTL strings without comments
We store FTL translations as serialized FTL messages with removed
comments (which are also part of the AST). Pontoon's FTL parser removes
comments from translations, and keeps them in source strings.

To detect whether a file contains translations or source strings, it
checks if the source resource has been provided. This patch makes sure
that is the case on file load, which means translations will be
imported without comments.