Add a new approve_translation view.
This creates a new view dedicated to approving translations, in order to not have to use update_translation, which requires a lot of parameters for just changing a status. It also refreshes History data from the server, to make sure that the data is correct instead of trying to fix it locally.