Add a Locales tab that shows translations for other locales.
This reworks the way data is refreshed for the History and Locales components, because those are not always mounted inside react-tabs. It introduces flow types in the API, and we should make use of that more in the future. I also adds support for a plural_form parameter to the other_locales view.