Fix bug 1513104
Add a separate Dockerfile.prod that can be used to create an image to
share on DockerHub. It starts from Alpine Node 10 to save on build
time and then adds the required packages for running the django app.
It is mostly the same as the dev Dockerfile, and they should be
consolidated later. The big departure from the Heroku approach is
not assuming you want a file with your env vars, but that they are
instead provided at container start time.