Fixes bug 1486509 - Add search and status filters to Translate.Next. (#1136)
The input allows a user to search for a string in the list of results. There is a delay before the search takes place, as to allow the user to type a full word without sending a query per letter.

Filters are simple links that change the URL and cause the entities list to refresh. This also adds a stats module to store counts coming with the entities.

All URL changes are now in the navigation module, meaning less logic is in the components. All actions changing the URL now keep other parameters intact, except the entity for reasons explained in the code.