Fix bug 1397377: Restore ability to delete translations
Translation can only be deleted, if it's rejected. That way we make the
Delete action of secondary importance compared to the first-rate review
actions - Approve & Reject. Hence, if you want to delete a non-rejected
translation, you'll need to make two clicks: first to reject and then to
delete a translation - the second click acting like a sort of confirmation step.